+++ package.json
@@ -11,6 +11,8 @@
   },
   "license": "ISC",
   "dependencies": {
+    "@google/maps": "^0.5.5",
+    "axios": "^0.19.0",
     "bcrypt": "^3.0.6",
     "cookie-parser": "^1.4.4",
     "dotenv": "^6.2.0",
