+++ server.js
@@ -43,8 +43,8 @@ app.engine(
 app.set("view engine", "handlebars");
 
 // Routes
-require("./routes/apiRoutes")(app);
-require("./routes/htmlRoutes")(app);
+require("./routes/apiRoutes");
+require("./routes/loginRoutes")(app);
 
 var syncOptions = { force: false };
